Temporary fix for stm32_delay.
Previously, it was using HAL_Delay which uses Systick. But the interrupt handler for
systick has been overwritten by FreeRTOS. So, the HAL_Delay was fundamentally not working.
We should definitely look into improving this with actualy timers but this is fine for now.
The only place this is used in blinking an LED and as a quiet delay between
buzzes when there is an error in sensor self testing.